钱包操作 异步回调

+++ aupay-wallet/src/main/java/com/aucloud/aupay/wallet/orm/service/WalletCollectTaskRecordService.java
@@ -21,7 +21,7 @@ import java.util.UUID;
 @Service
 public class WalletCollectTaskRecordService extends ServiceImpl<WalletCollectTaskRecordMapper, WalletCollectTaskRecord> implements IService<WalletCollectTaskRecord> {
 
-    public WalletCollectTaskRecord generateTaskRecord(Integer configId, Integer dependTaskId) {
+    public WalletCollectTaskRecord generateTaskRecord(Integer configId, Long dependTaskId) {
         String tradeNo = UUID.randomUUID().toString();
         WalletCollectTaskRecord taskRecord = new WalletCollectTaskRecord();
         taskRecord.setConfigId(configId);
